upgrade connection error log from warn to error
We had a connectivity issue between the client and ld-relay due to some
misconfiguration. It resulted in apps not being able to connect to
LaunchDarkly for a long time.

It took a while to realize that the SDK actually logged a warning. It'd
be nice if this was an error.