chore: refactoring evaluation responsibilities and adding EvaluatorInterface (#428)
This PR is purely refactoring, there should be no behavioral changes.

**Requirements**

- [x] I have added test coverage for new or changed functionality

- [x] I have followed the repository's [pull request submission
guidelines](../blob/master/CONTRIBUTING.md#submitting-pull-requests)
- [x] I have validated my changes against all supported platform
versions

**Related issues**

SC-236793

**Describe the solution you've provided**

Responsibility for handling determining evaluation results was divided
across LDClient and Evaluator classes. This PR pushes the responsibility
out of the LDClient and into the Evaluator to create a cleaner division.
The interface introduced sets up the next PR in this series which adds
the EvaluatorWithHooks class that will implement calling the evaluation
hooks at the correct time.

**Describe alternatives you've considered**

I considered not doing this refactoring, but that would end up with hook
invocation sprinkled in a few spots and more evaluation functions. We
already had functions called "evaluate" in 2 different classes, so it
would only get worse with time.