feat: adds evaluation hooks support (#429)
**Requirements**

- [x] I have added test coverage for new or changed functionality
- [x] I have followed the repository's [pull request submission
guidelines](../blob/master/CONTRIBUTING.md#submitting-pull-requests)
- [ ] I have validated my changes against all supported platform
versions
I have not personally done this locally, but the CI tests do compile for
multiple Java versions. The changes made in this PR are not expected to
have any impact on different platforms.

**Related issues**

SC-236793

**Describe the solution you've provided**

Implemented evaluation hooks as a decorator of the EvaluatorInterface.
Added configuration / builders related to setting up the hooks with the
SDK during initialization.